rewritten, works on 10.8 & 10.9
For some reason extraction and compression wasn't working on beta versions. Rewrote script not to use tmp to extract, and instead recompresses mounted /ram/initramfs/memory/bundles instead for more limited memory machines.

Since i can't add files in that way, a  new 01-lzo.sb is created and added to the mix, adding savechanges_lzo and dir2sb_lzo, as well as touch-to-tap, which you can delete to make persistant if you want.